+++ stdlib/public/SwiftShims/Random.h
@@ -1,0 +1,43 @@
+//===--- Random.h - Wrapper for the OS random number API --------*- C++ -*-===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2018 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+//
+//  A wrapper around the host OS's secure random number generator.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef SWIFT_STDLIB_SHIMS_RANDOM_H
+#define SWIFT_STDLIB_SHIMS_RANDOM_H
+
+#include "SwiftStddef.h"
+#include "Visibility.h"
+
+#if __has_feature(nullability)
+#pragma clang assume_nonnull begin
+#endif
+
+#ifdef __cplusplus
+namespace swift { extern "C" {
+#endif
+
+SWIFT_RUNTIME_STDLIB_SPI
+void _swift_stdlib_random(void *buf, __swift_size_t nbytes);
+
+#ifdef __cplusplus
+}} // extern "C", namespace swift
+#endif
+
+#if __has_feature(nullability)
+#pragma clang assume_nonnull end
+#endif
+
+#endif // SWIFT_STDLIB_SHIMS_RANDOM_H
+
