+++ stdlib/public/SwiftShims/module.modulemap
@@ -6,6 +6,7 @@ module SwiftShims {
   header "HeapObject.h"
   header "KeyPath.h"
   header "LibcShims.h"
+  header "Random.h"
   header "RefCount.h"
   header "RuntimeShims.h"
   header "RuntimeStubs.h"
