+++ stdlib/public/stubs/LibcShims.cpp
@@ -19,8 +19,6 @@
 #include <io.h>
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
-#include <Bcrypt.h>
-#pragma comment(lib, "Bcrypt.lib")
 #else
 #include <semaphore.h>
 #include <sys/ioctl.h>
@@ -34,13 +32,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#if __has_include(<sys/random.h>)
-#include <sys/random.h>
-#endif
-#include <sys/stat.h>
-#if __has_include(<sys/syscall.h>)
-#include <sys/syscall.h>
-#endif
 #include <sys/types.h>
 #include <type_traits>
 
@@ -48,7 +39,6 @@
 #include "swift/Basic/Lazy.h"
 #include "swift/Runtime/Config.h"
 #include "swift/Runtime/Debug.h"
-#include "swift/Runtime/Mutex.h"
 #include "../SwiftShims/LibcShims.h"
 
 using namespace swift;
@@ -235,84 +225,4 @@ size_t swift::_swift_stdlib_malloc_size(const void *ptr) {
 }
 #else
 #error No malloc_size analog known for this platform/libc.
-#endif
-
-// _swift_stdlib_random
-//
-// Should the implementation of this function add a new platform/change for a
-// platform, make sure to also update the documentation regarding platform
-// implementation of this function.
-// This can be found at: /docs/Random.md
-
-#if defined(__APPLE__)
-
-SWIFT_RUNTIME_STDLIB_SPI
-void swift::_swift_stdlib_random(void *buf, __swift_size_t nbytes) {
-  arc4random_buf(buf, nbytes);
-}
-
-#elif defined(_WIN32) && !defined(__CYGWIN__)
-#warning TODO: Test _swift_stdlib_random on Windows
-
-SWIFT_RUNTIME_STDLIB_SPI
-void swift::_swift_stdlib_random(void *buf, __swift_size_t nbytes) {
-  NTSTATUS status = BCryptGenRandom(nullptr,
-                                    static_cast<PUCHAR>(buf),
-                                    static_cast<ULONG>(nbytes),
-                                    BCRYPT_USE_SYSTEM_PREFERRED_RNG);
-  if (!BCRYPT_SUCCESS(status)) {
-    fatalError(0, "Fatal error: 0x%.8X in '%s'\n", status, __func__);
-  }
-}
-
-#else
-
-#undef  WHILE_EINTR
-#define WHILE_EINTR(expression) ({                                             \
-  decltype(expression) result = -1;                                            \
-  do { result = (expression); } while (result == -1 && errno == EINTR);        \
-  result;                                                                      \
-})
-
-SWIFT_RUNTIME_STDLIB_SPI
-void swift::_swift_stdlib_random(void *buf, __swift_size_t nbytes) {
-  while (nbytes > 0) {
-    __swift_ssize_t actual_nbytes = -1;
-
-#if defined(__NR_getrandom)
-    static const bool getrandom_available =
-      !(syscall(__NR_getrandom, nullptr, 0, 0) == -1 && errno == ENOSYS);
-  
-    if (getrandom_available) {
-      actual_nbytes = WHILE_EINTR(syscall(__NR_getrandom, buf, nbytes, 0));
-    }
-#elif __has_include(<sys/random.h>) && (defined(__CYGWIN__) || defined(__Fuchsia__))
-    __swift_size_t getentropy_nbytes = std::min(nbytes, __swift_size_t{256});
-    
-    if (0 == getentropy(buf, getentropy_nbytes)) {
-      actual_nbytes = getentropy_nbytes;
-    }
-#endif
-
-    if (actual_nbytes == -1) {
-      static const int fd = 
-        WHILE_EINTR(_swift_stdlib_open("/dev/urandom", O_RDONLY | O_CLOEXEC, 0));
-        
-      if (fd != -1) {
-        static StaticMutex mutex;
-        mutex.withLock([&] {
-          actual_nbytes = WHILE_EINTR(_swift_stdlib_read(fd, buf, nbytes));
-        });
-      }
-    }
-    
-    if (actual_nbytes == -1) {
-      fatalError(0, "Fatal error: %d in '%s'\n", errno, __func__);
-    }
-    
-    buf = static_cast<uint8_t *>(buf) + actual_nbytes;
-    nbytes -= actual_nbytes;
-  }
-}
-
 #endif
